refactor code - function
