+++ alllatsian/confluence_service_handle.py
@@ -1,8 +1,9 @@
 import re
 from atlassian import Confluence
 from utils.constant import jira_api_token
-from alllatsian import jira_service_handle
+from alllatsian import jira_service_handle, genarate_plan
 from utils import parser_content
+from data.data_app import lstUserStory, lstTask
 
 confluence = Confluence(
     url='https://bidv-ba-assistant317.atlassian.net/wiki',
@@ -27,4 +28,15 @@ def agent_gen_estimate_doc(promt):
     jira_service_handle.attach_link_confluence_to_task()
 
 
-# "https://bidv-ba-assistant317.atlassian.net/wiki/spaces/BAAI/pages/8028200/Page+Gen+12"
+def create_table_est_for_doc_step():
+    lst_header = []
+    table_body = ""
+    for i in range(len(lstTask)):
+        item_task = lstTask[i]
+        if item_task.user_story not in lst_header:
+            table_body = table_body + genarate_plan.generate_row_header(i, item_task)
+            lst_header.append(item_task.user_story)
+        else:
+            table_body = table_body + genarate_plan.generate_row_normal(i, item_task)
+
+    agent_gen_estimate_doc(genarate_plan.source_html_plan_doc(table_body))
