+++ utils/constant.py
@@ -1,6 +1,8 @@
 model_config = "gpt-4o-mini"
 user_config = "user"
-jira_api_token = 'api_key'
+jira_api_token = 'jira_api_token'
+open_api_key = 'open_api_key'
+
 
 html_source = """
 <!DOCTYPE html>
